ECA Database
A (here empty) database storing which credits every student in Royal
West Academy completed in each year. It colour codes the entries, and
has several additional automated features. Each student was required to
complete credits in three of four categories, and was not allowed to
miss the same category two years in a row. PLEASE ENABLE MACROS